Remove old serializer.
