+++ core/containers/array.h
@@ -5,7 +5,6 @@
 #include "core/formatter.h"
 #include "core/reflect.h"
 #include "core/memory/memory.h"
-#include "core/serialization/serializer.h"
 
 #include <type_traits>
 #include <initializer_list>
@@ -283,26 +282,6 @@ destroy(Array<T> &self)
 			destroy(self[i]);
 	}
 	array_deinit(self);
-}
-
-template <typename T>
-inline static void
-serialize(Serializer *serializer, const Array<T> &self)
-{
-	serialize(serializer, self.count);
-	for (u64 i = 0; i < self.count; ++i)
-		serialize(serializer, self[i]);
-}
-
-template <typename T>
-inline static void
-deserialize(Serializer *serializer, Array<T> &self)
-{
-	u64 count = 0;
-	deserialize(serializer, count);
-	array_resize(self, count);
-	for (u64 i = 0; i < count; ++i)
-		deserialize(serializer, self[i]);
 }
 
 template <typename T>
