+++ test/CMakeLists.txt
@@ -1,6 +1,4 @@
 set(HEADER_FILES
-    src/bin_serializer.h
-    src/jsn_serializer.h
 )
 
 set(SOURCE_FILES
